refactor: split graph
